requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
Click==7.0
dnspython==1.16.0
eventlet==0.24.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-SocketIO==3.3.2
greenlet==0.4.15
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.5.0
MarkupSafe==1.1.1
monotonic==1.5
PyJWT==1.7.1
PyMySQL==0.9.3
python-engineio==3.5.1
python-socketio==3.1.2
pytz==2019.1
redis==3.2.1
six==1.12.0
SQLAlchemy==1.3.2
vine==1.3.0
Werkzeug==0.15.2
